# backend/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Farmasi Queue API",
    "version": "1.0.0",
    "description": "API documentation for Antrian Farmasi system"
  },
  "host": "192.168.6.106:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/queue/GMCB": {
      "post": {
        "tags": [
          "Queue"
        ],
        "summary": "Create Antrian GMCB",
        "description": "Create Antrian GMCB",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "medicineType": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "origin": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/create-farmasi/GMCB": {
      "post": {
        "tags": [
          "Queue"
        ],
        "summary": "Create Antrian GMCB",
        "description": "Create Antrian GMCB",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "medicineType": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "origin": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/GMCB-appointment/": {
      "post": {
        "tags": [
          "GMCB"
        ],
        "summary": "Create a new appointment",
        "description": "Creates a new appointment",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "lokasi": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "GMCB"
        ],
        "summary": "Get All Appointment",
        "description": "Get All Appointment",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/get-list": {
      "post": {
        "tags": [
          "GMCB"
        ],
        "summary": "Get Farmasi List",
        "description": "Get Farmasi List",
        "parameters": [
          {
            "name": "x-cons-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-timestamp",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-signature",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "payload": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/antrian": {
      "get": {
        "tags": [
          "GMCB"
        ],
        "summary": "Get Antrian",
        "description": "Get Antrian",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/{NOP}": {
      "get": {
        "tags": [
          "GMCB"
        ],
        "summary": "Get By NOP",
        "description": "Get By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "GMCB"
        ],
        "summary": "Delete By NOP",
        "description": "Delete By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/{NOP}/phone_number": {
      "put": {
        "tags": [
          "GMCB"
        ],
        "summary": "Update By NOP",
        "description": "Update By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "phone_number": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/update_status/{NOP}": {
      "patch": {
        "tags": [
          "GMCB"
        ],
        "summary": "Update By NOP",
        "description": "Update By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/update/{NOP}": {
      "patch": {
        "tags": [
          "GMCB"
        ],
        "summary": "Update By NOP",
        "description": "Update By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "queue_number": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/total_medicine": {
      "patch": {
        "tags": [
          "GMCB"
        ],
        "summary": "Update total medicine by NOP",
        "description": "Update total medicine by NOP",
        "parameters": [
          {
            "name": "x-cons-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-timestamp",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-signature",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "NOP": {
                  "example": "any"
                },
                "total_medicine": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/type/{NOP}": {
      "put": {
        "tags": [
          "GMCB"
        ],
        "summary": "Update By NOP",
        "description": "Update By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status_medicine": {
                  "example": "any"
                },
                "farmasi_queue_number": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/{NOP}/status_medicine": {
      "patch": {
        "tags": [
          "GMCB"
        ],
        "summary": "Update By NOP",
        "description": "Update By NOP",
        "parameters": [
          {
            "name": "NOP",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status_medicine": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/by-date/{date}/{category}": {
      "get": {
        "tags": [
          "GMCB"
        ],
        "summary": "Get By Date",
        "description": "Get All Task By Date",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/GMCB-appointment/{category}": {
      "get": {
        "tags": [
          "GMCB"
        ],
        "summary": "Get By Category",
        "description": "Get By Category",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}